Added gitgrep and gitlog
Some bash scripts I find useful when using and perusing git repos.

gitgrep: A wrapper for git grep which breaks the output up per file and
provides 3 lines of context before and after the match.

gitlog: A wrapper for git log which provides a colourful tree version of
the log showing the title of each commit, the committer's name and how
long ago the commit occurred.